+++ apps/maestro/src/features/AddErc20/hooks/useDeployAndRegisterRemoteStandardizedTokenMutation.ts
@@ -165,13 +165,19 @@ export function useDeployAndRegisterRemoteStandardizedTokenMutation(
   useEffect(
     () => {
       if (recordDeploymentArgs) {
-        recordDeploymentAsync(recordDeploymentArgs).then(() => {
-          onStatusUpdate({
-            type: "deployed",
-            tokenAddress: recordDeploymentArgs.tokenAddress,
-            txHash: recordDeploymentArgs.deploymentTxHash,
+        recordDeploymentAsync(recordDeploymentArgs)
+          .then(() => {
+            onStatusUpdate({
+              type: "deployed",
+              tokenAddress: recordDeploymentArgs.tokenAddress,
+              txHash: recordDeploymentArgs.deploymentTxHash,
+            });
+          })
+          .catch(() => {
+            onStatusUpdate({
+              type: "idle",
+            });
           });
-        });
       }
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps