+++ apps/maestro/src/ui/compounds/ConnectWalletButton/ConnectWalletButton.spec.tsx
@@ -1,46 +1,78 @@
-import { render, waitFor } from "@testing-library/react";
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+
+import React from "react";
+
+import { fireEvent, render, waitFor } from "@testing-library/react";
 import { vi } from "vitest";
 
 import ConnectWalletButton from "./ConnectWalletButton";
 
-const mocks = {
-  useWeb3Modal: vi.fn(),
-};
+const mockOpen = vi.fn();
 
-vi.doMock("@web3modal/react", () => ({
-  useWeb3Modal: mocks.useWeb3Modal,
+vi.mock("@web3modal/react", () => ({
+  useWeb3Modal: vi.fn().mockImplementation(() => ({
+    open: mockOpen,
+  })),
 }));
 
 describe("ConnectWalletButton", () => {
-  it("should render correctly", async () => {
-    mocks.useWeb3Modal.mockReturnValue({
-      open: vi.fn(),
-    });
-
-    const { container } = render(<ConnectWalletButton />);
-
-    expect(container).toMatchInlineSnapshot(`
-      <div>
-        <button
-          class="btn btn-sm btn-primary"
-          data-testid="connect-button"
-          type="button"
-        >
-          Connect Wallet
-        </button>
-      </div>
-    `);
+  afterEach(() => {
+    vi.clearAllMocks();
   });
 
-  it("should call open on click", async () => {
-    const open = vi.fn();
+  it("renders the button with default text 'Connect Wallet'", () => {
+    const { getByText } = render(<ConnectWalletButton />);
+    expect(getByText("Connect Wallet")).toBeInTheDocument();
+  });
 
-    mocks.useWeb3Modal.mockReturnValue({ open });
-
+  it("triggers the open function from useWeb3Modal when clicked", async () => {
     const { getByTestId } = render(<ConnectWalletButton />);
+    const button = getByTestId("connect-button");
 
-    getByTestId("connect-button").click();
+    fireEvent.click(button);
 
-    waitFor(() => expect(open).toHaveBeenCalledTimes(1));
+    await waitFor(() => {
+      expect(mockOpen).toBeCalledTimes(1);
+    });
+  });
+
+  it("renders the button with default size 'sm'", () => {
+    const { getByTestId } = render(<ConnectWalletButton />);
+    const button = getByTestId("connect-button");
+
+    expect(button).toHaveClass("btn-sm");
+  });
+
+  it("renders the button with default variant 'primary'", () => {
+    const { getByTestId } = render(<ConnectWalletButton />);
+    const button = getByTestId("connect-button");
+
+    expect(button).toHaveClass("btn-primary");
+  });
+
+  it("renders the button with custom props", () => {
+    const { getByText } = render(
+      <ConnectWalletButton size="lg" variant="secondary">
+        Custom Connect
+      </ConnectWalletButton>
+    );
+    const button = getByText("Custom Connect");
+
+    expect(button).toHaveClass("btn-lg");
+    expect(button).toHaveClass("btn-secondary");
+  });
+
+  it("supports forwarding refs", async () => {
+    const ref = React.createRef<HTMLButtonElement>();
+    render(<ConnectWalletButton ref={ref} />);
+    expect(ref.current).toBeInstanceOf(HTMLButtonElement);
+
+    if (ref.current) {
+      fireEvent.click(ref.current);
+    }
+
+    await waitFor(() => {
+      expect(mockOpen).toBeCalledTimes(1);
+    });
   });
 });