+++ packages/api/src/deposit-address/index.ts
@@ -1,4 +1,3 @@
-export * from "./isomorphic";
-export * from "./browser";
-export * from "./node";
 export * from "./types";
+export * from "./isomorphic";
+export * from "./client";