+++ packages/transaction-recovery/src/add-gas/isomorphic.spec.ts
@@ -1,6 +1,7 @@
-import { createAxelarConfigNodeClient } from "@axelarjs/api/axelar-config/node";
-import { createAxelarQueryNodeClient } from "@axelarjs/api/axelar-query/node";
-import { createGMPNodeClient } from "@axelarjs/api/gmp/node";
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+import { createAxelarConfigClient } from "@axelarjs/api/axelar-config";
+import { createAxelarQueryClient } from "@axelarjs/api/axelar-query";
+import { createGMPClient } from "@axelarjs/api/gmp";
 import { ENVIRONMENTS } from "@axelarjs/core";
 
 import { DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
@@ -28,13 +29,21 @@ describe("addGas - (isomorphic)", () => {
   const mockGetSigningStargateClient = vi.fn(() =>
     Promise.resolve({
       signAndBroadcast: vi.fn().mockImplementation(mockSignAndBroadcast),
+      messages: {
+        ibcTransfer: {
+          transfer: {
+            signAndBroadcast: vi.fn().mockImplementation(mockSignAndBroadcast),
+          },
+        },
+      },
     })
   );
 
   const DEFAULT_ADD_GAS_DEPENDENCIES: AddGasDependencies = {
-    axelarQueryClient: createAxelarQueryNodeClient(ENVIRONMENTS.testnet),
-    configClient: createAxelarConfigNodeClient(ENVIRONMENTS.testnet),
-    gmpClient: createGMPNodeClient(ENVIRONMENTS.testnet),
+    axelarQueryClient: createAxelarQueryClient(ENVIRONMENTS.testnet),
+    configClient: createAxelarConfigClient(ENVIRONMENTS.testnet),
+    gmpClient: createGMPClient(ENVIRONMENTS.testnet),
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     getSigningStargateClient: mockGetSigningStargateClient as any,
   };