+++ packages/ui/src/hooks/useIntervalAsync.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from "react";
 
-export type GenericFunction = (...args: any[]) => any;
+export type GenericFunction = (...args: unknown[]) => unknown;
 
 export const useIntervalAsync = <TCallback extends GenericFunction>(
   callback: TCallback,
@@ -16,9 +16,11 @@ export const useIntervalAsync = <TCallback extends GenericFunction>(
   useEffect(() => {
     const tick = () => savedCallback.current();
     if (typeof delay === "number") {
-      //@ts-ignore
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       intervalRef.current = window.setInterval(tick, delay);
-      //@ts-ignore
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       return () => window.clearInterval(intervalRef.current);
     }
   }, [delay]);