+++ packages/cosmos/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @axelarjs/cosmos
 
+## 0.2.5
+
+### Patch Changes
+
+- [`4b8adb6f`](https://github.com/axelarnetwork/axelarjs/commit/4b8adb6f18b69745aad3045519d870e880ec226a) Thanks [@alanrsoares](https://github.com/alanrsoares)! - - Migrage api clients to use minimalistic isomorphic http client based on `isomorphic-unfetc`
+  - upgrade dependencies
+- Updated dependencies [[`4b8adb6f`](https://github.com/axelarnetwork/axelarjs/commit/4b8adb6f18b69745aad3045519d870e880ec226a)]:
+  - @axelarjs/config@0.1.7
+  - @axelarjs/proto@0.34.7
+  - @axelarjs/utils@0.1.8
+
+## 0.2.4
+
+### Patch Changes
+
+- [#59](https://github.com/axelarnetwork/axelarjs/pull/59) [`72df92de`](https://github.com/axelarnetwork/axelarjs/commit/72df92de09c0d41156708ce3f93263c039370675) Thanks [@alanrsoares](https://github.com/alanrsoares)! - Introduce typesafe native cosmos messages via AxelarSigningStargateClient
+
+## 0.2.3
+
+### Patch Changes
+
+- [`a7457a7b`](https://github.com/axelarnetwork/axelarjs/commit/a7457a7b5ad7d0b5bcc1dfcf483dc335ea61cb86) Thanks [@canhtrinh](https://github.com/canhtrinh)! - update build settings for deposit-address module
+
+- Updated dependencies [[`a7457a7b`](https://github.com/axelarnetwork/axelarjs/commit/a7457a7b5ad7d0b5bcc1dfcf483dc335ea61cb86)]:
+  - @axelarjs/config@0.1.6
+  - @axelarjs/proto@0.34.6
+  - @axelarjs/utils@0.1.7
+
+## 0.2.2
+
+### Patch Changes
+
+- [`59d3ff8f`](https://github.com/axelarnetwork/axelarjs/commit/59d3ff8fa77fefe3639a8fb1f7cf3263162c360e) Thanks [@canhtrinh](https://github.com/canhtrinh)! - bumping all dependencies
+
+- Updated dependencies [[`59d3ff8f`](https://github.com/axelarnetwork/axelarjs/commit/59d3ff8fa77fefe3639a8fb1f7cf3263162c360e)]:
+  - @axelarjs/config@0.1.5
+  - @axelarjs/proto@0.34.5
+  - @axelarjs/utils@0.1.6
+
+## 0.2.1
+
+### Patch Changes
+
+- [`c9efbb15`](https://github.com/axelarnetwork/axelarjs/commit/c9efbb1523ca6cf06d56865a0115cf214e16b1c7) Thanks [@alanrsoares](https://github.com/alanrsoares)! - Do not use path aliases in publishable packages
+
+- Updated dependencies [[`c9efbb15`](https://github.com/axelarnetwork/axelarjs/commit/c9efbb1523ca6cf06d56865a0115cf214e16b1c7)]:
+  - @axelarjs/proto@0.34.4
+  - @axelarjs/utils@0.1.5
+
+## 0.2.0
+
+### Minor Changes
+
+- [`f6a6c66b`](https://github.com/axelarnetwork/axelarjs/commit/f6a6c66bdd347b4e7453c07bf2360a97dbdaefde) Thanks [@alanrsoares](https://github.com/alanrsoares)! - Update @axelarjs/cosmos package structure
+
+## 0.1.6
+
+### Patch Changes
+
+- [`c2b0fa7c`](https://github.com/axelarnetwork/axelarjs/commit/c2b0fa7c3920102a30e3e6d205e5574586c47d98) Thanks [@alanrsoares](https://github.com/alanrsoares)! - - @axelarjs/cosmos: Implement AxelarSigningStargateClient
+  - @axelarjs/utils: Add `memoize` util to `functions` module
+  - @axelarjs/proto: Upgrade protobuf dependencies
+  - all: upgrade vitest
+- Updated dependencies [[`c2b0fa7c`](https://github.com/axelarnetwork/axelarjs/commit/c2b0fa7c3920102a30e3e6d205e5574586c47d98)]:
+  - @axelarjs/config@0.1.4
+  - @axelarjs/proto@0.34.3
+  - @axelarjs/utils@0.1.4
+
 ## 0.1.5
 
 ### Patch Changes